test: mimic code structure in unit tests (#735)
* chore: removed license from test files

* test: mimic code structure in unit tests
