+++ tests/unit/commands/openapi_generator/test_json_to_object.py
