+++ tests/unit/commands/openapi_generator/test_oas.py
