+++ tests/unit/commands/openapi_generator/test_object_to_json.py
